3.Collision 20.5.2021: Collisions work, for the most part
Resource collisions aren't working for some reason. It could have something to do with  the inheritance.

The bigger issue for us is that right now we can't check for, what object is colliding with us, making it impossible implement game mechanics correctly.

Need to return to this later today, need to go to shop, since tomorrow will be rainy. To be presice, it will be stormy day.