1. FSM 21.5.2021: FSM implemented.
It is not perfect, quite frankly this system is confusing, but at this point I have to make due with this. But at least managed to eliminate all the bugs in the system.

Next I need to make spawner for the resources and work on perfecting the game mechanics. But before that, I need to program the database system, since deadline for that is much closer than for AI programming.