4. FSM 22.7.2021: FSM system overhaul
Made a new version of this, not sure if it's any logical than the original or if it will allow multiple FSM entities to exist in the game without causing problems.

What I would really want to do is get this to work in a way that both Wolf and Kitten can use it and still get different behaviour.